requirements.txt
chromadb
openai
langchain[llms]
python-dotenv
numexpr

PyGithub
sentence_transformers
unstructured

fastapi
uvicorn[standard]

GitPython
